Some simplification of the API; use longs not doubles for counts, etc.

+++ metrics/api/src/main/java/io/helidon/metrics/api/CountAtBucket.java
@@ -47,5 +47,5 @@ public interface CountAtBucket extends Wrapped {
      *
      * @return observation count for the bucket
      */
-    double count();
+    long count();
 }
